# Optimize Activity::IsActivity helper method