--- a/libraries/Microsoft.Bot.Schema/ActivityEx.cs
+++ b/libraries/Microsoft.Bot.Schema/ActivityEx.cs
@@ -178,7 +178,36 @@ namespace Microsoft.Bot.Schema
         /// <summary>
         /// True if the Activity is of the specified activity type
         /// </summary>
-        protected bool IsActivity(string activity) { return string.Compare(this.Type?.Split('/').First(), activity, true) == 0; }
+        protected bool IsActivity(string activityType)
+        {
+            var type = this.Type;
+
+            // If there's no type set then we can't tell if it's the type they're looking for
+            if (type == null)
+            {
+                return false;
+            }
+
+            // Check if the full type value starts with the type they're looking for
+            var result = type.StartsWith(activityType, StringComparison.OrdinalIgnoreCase);
+
+            // If the full type value starts with the type they're looking for, then we need to check a little further to check if it's definitely the right type
+            if (result)
+            {
+                // If the lengths are equal, then it's the exact type they're looking for
+                result = type.Length == activityType.Length;
+
+                if (!result)
+                {
+                    // Finally, if the type is longer than the type they're looking for then we need to check if there's a / separator right after the type they're looking for
+                    result = type.Length > activityType.Length
+                                    &&
+                            type[activityType.Length] == '/';
+                }
+            }
+
+            return result;
+        }
 
         /// <summary>
         /// Return an IMessageActivity mask if this is a message activity