@cleemullins 2018-05-29
@drub0y This looks awesome. Could you just include a comment saying why this code is optimized, so that when I next look at it some years down the line I don't go "Why is this so complicated? This could be a single line of code!".