@drub0y 2018-05-30
Done. And will try to make sure to do this in the future as well as it makes a lot of sense.